AI Interviewer/requirements.txt
fastapi
uvicorn
opencv-python
pytesseract
pyaudio
numpy
torch
openai-whisper
mss